+++ lib/my_main_app.dart
@@ -1,3 +1,4 @@
+import 'package:device_preview/device_preview.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
@@ -29,6 +30,10 @@ class _MyMaterialApp extends StatelessWidget {
           ? kDarkTheme
           : kLightTheme,
       routes: kAppRoutingTable,
+      debugShowCheckedModeBanner: false,
+      useInheritedMediaQuery: true,
+      locale: DevicePreview.locale(context),
+      builder: DevicePreview.appBuilder,
     );
   }
 }