+++ lib/routes/growth_device_preview_ex.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class InAppReviewExample extends StatelessWidget {
+  const InAppReviewExample({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(); // TODO
+  }
+}